+++ front/src/components/shared/Header.vue
@@ -9,7 +9,7 @@
 
     <v-spacer></v-spacer>
 
-    <router-link v-show="!spScreen" :to="{ name: 'Sentences' }" class="mr-5 header-text text-grey-darken-4">
+    <router-link v-show="!spScreen" :to="{ name: 'NewsSentences' }" class="mr-5 header-text text-grey-darken-4">
       英文一覧
     </router-link>
 
@@ -196,7 +196,7 @@ function linkToMypage(): void{
 
 function linkToSentences(): void{
   drawer.value = false
-  router.push({ name: "Sentences" })
+  router.push({ name: "NewsSentences" })
 }
 
 function linkToMySentences(): void{
